check before rm

+++ v2.sh
@@ -138,6 +138,9 @@ install_v2ray() {
 }
 
 rm_v2ray() {
+  if [ ! -d "/usr/bin/v2ray" ] || [ ! -f "/usr/local/bin/tls-shunt-proxy" ]; then
+    return 1
+  fi
   ${sudoCmd} ${systemPackage} install curl -y
 
   # remove v2ray
@@ -191,7 +194,6 @@ generate_link() {
 }
 
 menu() {
-  clear
   colorEcho ${YELLOW} "v2Ray TCP+TLS+WEB automated script v0.1"
   colorEcho ${YELLOW} "author: phlinhng"
   echo ""
